pyproject.toml
[project]
name = "learn-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "authlib>=1.6.4",
    "bcrypt>=4.3.0",
    "bson>=0.5.10",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "ruff>=0.12.12",
    "scalar-fastapi>=1.3.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

# https://docs.astral.sh/ruff/rules/#flake8-bandit-s
[tool.ruff]
line-length = 120
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # flake8 errors
    "I", # isort errors
]
exclude=["alembic"]
